Adapting code to new video box class
